requirements.txt

streamlit
langchain
langchain-community
langchain-text-splitters
huggingface_hub
sentence-transformers
faiss-cpu
